Functional Defense
I just need to add the images and fix the weird jidder bug with dragable. there is some more over all structure stuff like getting it to know which one should add when there are multiple dropzones. I could assign them to a specific spot in the list and specific hp ui shown during game play that would be easiest