Added moving up enemy agression
there are a couple issues. one if there are 2 of the same enemies its hard to tell they swapped. two the pointer stays on the one that got moved back so I need to have it disable its pointer and enable the one that moved up if its in the right spot. also I wonder if the wait timers have updated