Created wave generation
there might also be a bug with agression. it might with a regular weapon make enemies un attackable. at the very least sometimes the last enemy alive is at the second position it happened with the odachi just now. I think its actually not a problem. the bigger problem is that the difficulty caclulation doesn't work because if the first enemy added to the wave always only compares the first number it has not the whole thing, so I need a better calculation, but hey its a start :D