started changes to make enemies less likley to repeat
im not 100% sure I want this but the idea is after each wave I want it less likley they see repeat enemies. I think I should make it 50% chance the enemy is removed from list so you can still have enemies that are 2 or even 3 of the same. but anyway ima work on it after this final