Fixed gameplay flow issue
kept setting variable to false which I think occured after it was being set to true and this setting to false happened before the check that would continue the flow. CURRENT BUG: the armory is all fucked up! armory panel issues: when the panel is open it freees. when the button is disabled while the panel is active on camera it causes a red error